refactor: move the session routes out of index.ts (#548 step 2b-2)
The last and most entangled group: the sidebar listing, one session's detail,
the grid's attention poll, the tool timeline, and codex's own sessions. They
only became movable once step 1 made the state a module and step 2b-1 did the
same for the readers — what would have been eight injected functions is now one.

`freshenRosterTitle` is that one: re-titling a viewed session spawns a
summarizer, which belongs to the title machinery index.ts still owns.

Two more pieces move to where they were already needed rather than being
injected: translationWorkerIds to the registry (per-session state that routes,
hooks and the translation worker all read) and hasErrnoCode to errors.ts.

Lint caught the mount at 97 lines. Rather than raise the ceiling, each route is
now a named handler — sessionDetail, activitySnapshot, toolTimeline,
sessionList, codexSessionList — and the mount is five registrations. Standalone
handlers lose Express's path-literal param inference, so sessionDetail takes
Request<{ id: string }>, the same shape backends/feeds.ts uses.

Verified as before, with the invalid-input paths included this time since the
400s live in the moved code: main and this branch booted side by side, nine
requests compared (five happy, three rejected ids, one unscoped listing) —
identical. index.ts additions are the mount call and its imports.

index.ts: 2514 → 2359 lines.

Refs #548
